Renaming package to terra-navigation-layout
