+++ packages/terra-framework-site/package.json
@@ -44,7 +44,7 @@
     "terra-list": "^1.12.1",
     "terra-markdown": "^1.9.1",
     "terra-menu": "^1.6.0",
-    "terra-navigation": "^0.x",
+    "terra-navigation-layout": "^0.x",
     "terra-props-table": "^1.10.1",
     "terra-slide-panel": "^1.11.1",
     "terra-theme-provider": "^1.6.1"
