+++ packages/terra-framework-site/src/examples/navigation-layout/Index.jsx
@@ -2,15 +2,15 @@
 import React from 'react';
 import PropsTable from 'terra-props-table';
 import Markdown from 'terra-markdown';
-import ReadMe from 'terra-navigation/docs/README.md';
-import { version } from 'terra-navigation/package.json';
+import ReadMe from 'terra-navigation-layout/docs/README.md';
+import { version } from 'terra-navigation-layout/package.json';
 
 // Component Source
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
-import NavigationLayoutSrc from '!raw-loader!terra-navigation/src/NavigationLayout.jsx';
+import NavigationLayoutSrc from '!raw-loader!terra-navigation-layout/src/NavigationLayout.jsx';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 
-const NavigationExamples = () => (
+const NavigationLayoutExamples = () => (
   <div>
     <div id="version">Version: {version}</div>
     <Markdown id="readme" src={ReadMe} />
@@ -18,4 +18,4 @@ const NavigationExamples = () => (
   </div>
 );
 
-export default NavigationExamples;
+export default NavigationLayoutExamples;
