+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -1,0 +1,6 @@
+Changelog
+=========
+
+Unreleased
+----------
+* Initial release of terra-navigation-layout
