+++ packages/terra-navigation-layout/README.md
@@ -1,13 +1,14 @@
-# Terra Clinical NavigationSecondary
+# Terra Navigation Layout
 
+The Terra Navigation Layout package includes a variety of components and utilities to provide configuration-based, `react-router`-driven navigation to Terra applications.
 
-[![NPM version](http://img.shields.io/npm/v/terra-navigation.svg)](https://www.npmjs.org/package/terra-navigation)
+[![NPM version](http://img.shields.io/npm/v/terra-navigation-layout.svg)](https://www.npmjs.org/package/terra-navigation-layout)
 [![Build Status](https://travis-ci.org/cerner/terra-framework.svg?branch=master)](https://travis-ci.org/cerner/terra-framework)
 
 This component renders secondary navigation items with react router.
 
 - [Getting Started](#getting-started)
-- [Documentation](https://github.com/cerner/terra-framework/tree/master/packages/terra-navigation/docs)
+- [Documentation](https://github.com/cerner/terra-framework/tree/master/packages/terra-navigation-layout/docs)
 - [LICENSE](#license)
 
 ## Getting Started
