+++ packages/terra-navigation-layout/src/NavigationLayoutContent.scss
