+++ packages/terra-navigation-layout/src/configurationPropTypes.js
