+++ packages/terra-navigation-layout/src/routing/RoutingStackDelegate.js
