--- packages/terra-navigation/CHANGELOG.md
@@ -1,6 +1,0 @@
-Changelog
-=========
-
-Unreleased
-----------
-* Initial release of terra-navigation
